rename some engine files and fix puzzles not being included
