+++ client/src/scenes/ReplayBrowser.lua
@@ -3,7 +3,7 @@ local sceneManager = require("client.src.scenes.sceneManager")
 local input = require("common.lib.inputManager")
 local GraphicsUtil = require("client.src.graphics.graphics_util")
 local fileUtils = require("client.src.FileUtils")
-local Replay = require("common.engine.replay")
+local Replay = require("common.engine.Replay")
 local class = require("common.lib.class")
 local GameModes = require("common.engine.GameModes")
 local ReplayGame = require("client.src.scenes.ReplayGame")
