+++ common/engine/Match.lua
@@ -11,7 +11,7 @@ local ChallengeModePlayer = require("client.src.ChallengeModePlayer")
 local logger = require("common.lib.logger")
 local tableUtils = require("common.lib.tableUtils")
 local GameModes = require("common.engine.GameModes")
-local Replay = require("common.engine.replay")
+local Replay = require("common.engine.Replay")
 local Signal = require("common.lib.signal")
 local SimulatedStack = require("common.engine.SimulatedStack")
 local consts = require("common.engine.consts")
