+++ common/engine/PanelGenerator.lua
